move logic to localconf.js
